+++ backend/app/models/dict_data.py
@@ -1,10 +1,10 @@
 from datetime import datetime
-from sqlalchemy import Column, BigInteger, String, SmallInteger, Text, DateTime
+from sqlalchemy import Integer, Column, BigInteger, String, SmallInteger, Text, DateTime
 from app.core.database import Base
 
 class DictType(Base):
     __tablename__ = "sys_dict_type"
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     dict_name = Column(String(100), nullable=False)
     dict_type = Column(String(100), unique=True, nullable=False)
     status = Column(SmallInteger, default=1)
@@ -14,7 +14,7 @@ class DictType(Base):
 
 class DictData(Base):
     __tablename__ = "sys_dict_data"
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     dict_type = Column(String(100), nullable=False, index=True)
     dict_label = Column(String(100), nullable=False)
     dict_value = Column(String(100), nullable=False)
