+++ backend/app/models/knowledge.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from sqlalchemy import Column, BigInteger, String, Text, DateTime
+from sqlalchemy import Integer, Column, BigInteger, String, Text, DateTime
 from app.core.database import Base
 
 try:
@@ -10,7 +10,7 @@ except ImportError:
 
 class Knowledge(Base):
     __tablename__ = "knowledge"
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     title = Column(String(200), nullable=False)
     content = Column(Text, nullable=False)
     category = Column(String(50), nullable=False)
