+++ backend/app/models/relationship.py
@@ -1,10 +1,10 @@
 from datetime import datetime
-from sqlalchemy import Column, BigInteger, String, SmallInteger, Text, DateTime, ForeignKey
+from sqlalchemy import Integer, Column, BigInteger, String, SmallInteger, Text, DateTime, ForeignKey
 from app.core.database import Base
 
 class Relationship(Base):
     __tablename__ = "relationships"
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     customer_id = Column(BigInteger, ForeignKey("organizations.id"), nullable=False)
     contact_name = Column(String(100), nullable=False)
     contact_role = Column(String(50))
